client-followup-autopilot-cambios-V6/requirements.txt
python-dotenv
requests
anthropic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
slack-sdk
schedule
python-dateutil
pytz
